Rename: styles/폴더 생성, 경로설정 및 전역 스타일 파일 이동
