+++ src/main.tsx
@@ -3,8 +3,8 @@ import ReactDOM from 'react-dom/client'
 import { ThemeProvider } from 'styled-components'
 import { RouterProvider } from 'react-router-dom'
 
-import { basicTheme } from '@layouts/theme.ts'
-import { GlobalStyle } from '@layouts/GlobalStyle.ts'
+import { basicTheme } from '@/styles/theme.ts'
+import { GlobalStyle } from '@/styles/GlobalStyle.ts'
 import router from './router.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
